rest_framework_hmac/example_project/requirements.txt
Django==2.0.3
django-rest-framework-hmac
djangorestframework==3.7.7
